style: create separate mock folder for analytics manager
