+++ packages/analytics_manager/test/mocks/mock_analytics_manager.dart
@@ -1,0 +1,22 @@
+// ignore_for_file: no-empty-block
+
+import 'package:analytics_manager/analytics_manager.dart';
+
+final class MockAnalyticsManager extends AnalyticsManager {
+  const MockAnalyticsManager({super.logManager});
+
+  @override
+  Future<void> init() async {}
+
+  @override
+  void dispose() {}
+
+  @override
+  Future<void> resetAnalyticsData() async {}
+
+  @override
+  Future<void> enableAnalytics() async {}
+
+  @override
+  Future<void> disableAnalytics() async {}
+}
