+++ packages/analytics_manager/test/mocks/mock_firebase_analytics.dart
@@ -1,0 +1,4 @@
+import 'package:firebase_analytics/firebase_analytics.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockFirebaseAnalytics extends Mock implements FirebaseAnalytics {}
