+++ packages/analytics_manager/test/mocks/mock_log_manager.dart
@@ -1,0 +1,5 @@
+// Mock implementation of the LogManager for testing
+import 'package:log_manager/log_manager.dart';
+import 'package:mocktail/mocktail.dart';
+
+final class MockLogManager extends Mock implements LogManager {}
